Rename project and testing files to Rex
