+++ README.md
@@ -1,6 +1,6 @@
-# Duke project template
+# Hospitalsaurus ReX project template
 
-This is a project template for a greenfield Java project. It's named after the Java mascot _Duke_. Given below are instructions on how to use it.
+This is a project template for a greenfield Java project. Given below are instructions on how to use it.
 
 ## Setting up in Intellij
 
@@ -8,20 +8,22 @@ Prerequisites: JDK 11 (use the exact version), update Intellij to the most recen
 
 1. **Configure Intellij for JDK 11**, as described [here](https://se-education.org/guides/tutorials/intellijJdk.html).
 1. **Import the project _as a Gradle project_**, as described [here](https://se-education.org/guides/tutorials/intellijImportGradleProject.html).
-1. **Verify the set up**: After the importing is complete, locate the `src/main/java/seedu/duke/Duke.java` file, right-click it, and choose `Run Duke.main()`. If the setup is correct, you should see something like the below:
+1. **Verify the set up**: After the importing is complete, locate the `src/main/java/seedu/rex/Rex.java` file, right-click it, and choose `Run Rex.main()`. If the setup is correct, you should see something like the below:
    ```
-   > Task :compileJava
+   > Task :compileJava UP-TO-DATE
    > Task :processResources NO-SOURCE
-   > Task :classes
+   > Task :classes UP-TO-DATE
    
-   > Task :Duke.main()
+   > Task :Rex.main()
    Hello from
-    ____        _        
-   |  _ \ _   _| | _____ 
-   | | | | | | | |/ / _ \
-   | |_| | |_| |   <  __/
-   |____/ \__,_|_|\_\___|
-   
+     _    _                 _ _        _                                  _____     __   __
+    | |  | |               (_| |      | |                                |  __ \    \ \ / /
+    | |__| | ___  ___ _ __  _| |_ __ _| |___  __ _ _   _ _ __ _   _ ___  | |__) |___ \ V /
+    |  __  |/ _ \/ __| '_ \| | __/ _` | / __|/ _` | | | | '__| | | / __| |  _  // _ \ > <  
+    | |  | | (_) \__ | |_) | | || (_| | \__ | (_| | |_| | |  | |_| \__ \ | | \ |  __// . \
+    |_|  |_|\___/|___| .__/|_|\__\__,_|_|___/\__,_|\__,_|_|   \__,_|___/ |_|  \_\___/_/ \_\
+                     | |                                                                   
+                     |_|
    What is your name?
    ```
    Type some word and press enter to let the execution proceed to the end.
@@ -39,7 +41,7 @@ Prerequisites: JDK 11 (use the exact version), update Intellij to the most recen
 
 ### JUnit tests
 
-* A skeleton JUnit test (`src/test/java/seedu/duke/DukeTest.java`) is provided with this project template. 
+* A skeleton JUnit test (`src/test/java/seedu/rex/RexTest.java`) is provided with this project template. 
 * If you are new to JUnit, refer to the [JUnit Tutorial at se-education.org/guides](https://se-education.org/guides/tutorials/junit.html).
 
 ## Checkstyle
