+++ build.gradle
@@ -29,11 +29,11 @@ test {
 }
 
 application {
-    mainClassName = "seedu.duke.Duke"
+    mainClassName = "seedu.rex.Rex"
 }
 
 shadowJar {
-    archiveBaseName = "duke"
+    archiveBaseName = "rex"
     archiveClassifier = null
 }
