+++ src/test/java/seedu/rex/RexTest.java
@@ -1,10 +1,10 @@
-package seedu.duke;
+package seedu.rex;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-class DukeTest {
+class RexTest {
     @Test
     public void sampleTest() {
         assertTrue(true);
