+++ text-ui-test/EXPECTED.TXT
@@ -1,9 +1,11 @@
 Hello from
- ____        _        
-|  _ \ _   _| | _____ 
-| | | | | | | |/ / _ \
-| |_| | |_| |   <  __/
-|____/ \__,_|_|\_\___|
-
+  _    _                 _ _        _                                  _____     __   __
+ | |  | |               (_| |      | |                                |  __ \    \ \ / /
+ | |__| | ___  ___ _ __  _| |_ __ _| |___  __ _ _   _ _ __ _   _ ___  | |__) |___ \ V / 
+ |  __  |/ _ \/ __| '_ \| | __/ _` | / __|/ _` | | | | '__| | | / __| |  _  // _ \ > <  
+ | |  | | (_) \__ | |_) | | || (_| | \__ | (_| | |_| | |  | |_| \__ \ | | \ |  __// . \ 
+ |_|  |_|\___/|___| .__/|_|\__\__,_|_|___/\__,_|\__,_|_|   \__,_|___/ |_|  \_\___/_/ \_\
+                  | |                                                                   
+                  |_|                                                                   
 What is your name?
 Hello James Gosling
